integrates the proof for the inverse by Elie Ancelin
The inverse function takes an extra argument by comparison with the previous
version.  This extra argument is the rank of the most significant digit for
the input.  This serves as a witness that the input is non-zero.

The previous version of the proof relied on the function msd and its
specification, which basically assumed that every number was non-zero.